Added a plopfile to make it easier to write new serverside code.
