+++ package.json
@@ -25,22 +25,22 @@
     "chai": "^4.1.2",
     "chai-http": "^3.0.0",
     "husky": "^0.14.3",
+    "inquirer-directory": "^2.1.0",
+    "inquirer-file-path": "^1.0.0",
+    "inquirer-recursive": "0.0.5",
     "jest-localstorage-mock": "^2.2.0",
     "lint-staged": "^6.1.0",
     "nodemon": "^1.14.12",
     "pkg": "^4.3.0",
+    "plop": "^1.9.1",
     "prettier": "^1.10.2",
     "react-scripts": "1.1.0"
   },
   "dependencies": {
     "apollo-cache-inmemory": "^1.1.7",
-    "apollo-client": "^2.2.3",
-    "apollo-link-http": "^1.3.3",
     "apollo-client": "^2.2.2",
     "apollo-link-http": "^1.3.3",
     "apollo-link-ws": "^1.0.5",
-    "apollo-link-http": "^1.3.3",
-    "apollo-link-ws": "^1.0.4",
     "appdirectory": "^0.1.0",
     "body-parser": "^1.18.2",
     "bodymovin": "^4.13.0",
@@ -117,7 +117,8 @@
     "precommit": "lint-staged",
     "add": "all-contributors add",
     "generate": "all-contributors generate",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "plop": "plop"
   },
   "lint-staged": {
     "*.js": [
