+++ server/schema/mutations/actions.js
@@ -1,4 +1,5 @@
 export default `
 #Macro: Trigger Action (eg. Flash, Blackout, etc.)
 triggerAction(action: String!, simulatorId: ID!, stationId: String, clientId: ID, duration: Float): String
+something(one:String):String
 `;
